requirements.txt
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
cssselect==1.2.0
cssutils==2.11.1
Faker==36.1.1
idna==3.10
lxml==5.3.1
more-itertools==10.6.0
premailer==3.10.0
python-dotenv==1.0.1
requests==2.32.3
tzdata==2025.1
urllib3==2.3.0
yagmail==0.15.293
